+++ src/policy/policy.cpp
@@ -15,7 +15,7 @@
 
 #include <boost/foreach.hpp>
 // SYSCOIN need constant SYSCOIN_TX_VERSION
-#include "alias.h"
+#include "services/alias.h"
     /**
      * Check transaction inputs to mitigate two
      * potential denial-of-service attacks:
