+++ src/services/asset.cpp
